[IGNORE] Move DatasourceEditorForm to plugin-system pkg + some renaming for clarity & alignment between vars & ds CRUD
