+++ ui/plugin-system/src/components/DatasourceEditorForm/index.ts
@@ -11,16 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './CalculationSelector';
 export * from './DatasourceEditorForm';
-export * from './DatasourceSelect';
-export * from './LegendOptionsEditor';
-export * from './OptionsEditorRadios';
-export * from './OptionsEditorTabs';
-export * from './PanelSpecEditor';
-export * from './PluginEditor';
-export * from './PluginKindSelect';
-export * from './PluginRegistry';
-export * from './PluginSpecEditor';
-export * from './TimeSeriesQueryEditor';
-export * from './Variables';
